+++ ASP.NetCore7.ProjectMuseum/ProjectMuseum.Models/Museum.cs
@@ -4,5 +4,5 @@ public class Museum
 {
     public string Id { get; set; }
     public string Name { get; set; }
-    public int Money { get; set; }
+    public float Money { get; set; }
 }
