+++ ASP.NetCore7.ProjectMuseum/ProjectMuseum.Repositories/MuseumRepository/IMuseumRepository.cs
@@ -7,8 +7,8 @@ public interface IMuseumRepository
     Task<Museum> Insert(Museum museum);
     Task<Museum?> GetById(string id);
     Task<List<Museum>?> GetAll();
-    Task<int> GetMuseumBalance(string id);
-    Task<Museum> ReduceMuseumBalance(string id, int amount);
-    Task<Museum> AddToMuseumBalance(string id, int amount);
+    Task<float> GetMuseumBalance(string id);
+    Task<Museum> ReduceMuseumBalance(string id, float amount);
+    Task<Museum> AddToMuseumBalance(string id, float amount);
     Task<Museum?> Delete(string id);
 }
