+++ ASP.NetCore7.ProjectMuseum/ProjectMuseum.Services/MuseumService/IMuseumService.cs
@@ -4,7 +4,7 @@ namespace ProjectMuseum.Services.MuseumService;
 
 public interface IMuseumService
 {
-    Task<int> GetMuseumCurrentMoneyAmount(string id);
-    Task<Museum> AddToMuseumBalance(string id, int amount);
-    Task<Museum> ReduceMuseumBalance(string id, int amount);
+    Task<float> GetMuseumCurrentMoneyAmount(string id);
+    Task<Museum> AddToMuseumBalance(string id, float amount);
+    Task<Museum> ReduceMuseumBalance(string id, float amount);
 }
